error

+++ User/Motor/wheel.h
@@ -13,7 +13,7 @@ typedef struct WHEEL {
   uint8_t which;      // 电机编号 1-4
   int16_t trust;      // 电机推力
   int16_t cur_speed;  // 电机当前速度
-  int16_t tar_speed;  // 电机目标速度
+  float tar_speed;    // 电机目标速度
   int8_t dir;         // 电机方向
   PID wheel_pid;
 } WHEEL;
